+++ src/app/CliTools/Console/Command/Docker/AbstractCommand.php
@@ -4,6 +4,7 @@ namespace CliTools\Console\Command\Docker;
 
 /*
  * CliTools Command
+ * Copyright (C) 2016 WebDevOps.io
  * Copyright (C) 2015 Markus Blaschke <markus@familie-blaschke.net>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -54,7 +55,29 @@ abstract class AbstractCommand extends \CliTools\Console\Command\AbstractCommand
     }
 
     /**
-     * Execute docker command
+     * Find and get docker environment variable
+     *
+     * @param  string $containerName Container name
+     * @param  array  $envNameList   Environment variable (list)
+     *
+     * @return string|bool|null
+     */
+    protected function findAndGetDockerEnv($containerName, array $envNameList)
+    {
+        $ret = null;
+
+        foreach ($envNameList as $envName) {
+            if ($tmp = $this->getDockerEnv($containerName, $envName) ) {
+                $ret = $tmp;
+                break;
+            }
+        }
+
+        return $ret;
+    }
+
+    /**
+     * Get docker environment variable
      *
      * @param  string $containerName Container name
      * @param  string $envName       Environment variable