+++ src/app/CliTools/Console/Command/Docker/UpCommand.php
@@ -4,6 +4,7 @@ namespace CliTools\Console\Command\Docker;
 
 /*
  * CliTools Command
+ * Copyright (C) 2016 WebDevOps.io
  * Copyright (C) 2015 Markus Blaschke <markus@familie-blaschke.net>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -58,6 +59,8 @@ class UpCommand extends AbstractCommand
         $lastDockerPath = $this->getApplication()
                                ->getSettingsService()
                                ->get('docker.up.last');
+        $dockerComposeArgs = $this->getApplication()
+                                  ->getConfigValue('dockerCompose', 'up');
 
         if (!empty($dockerPath)) {
             $dockerPath = dirname($dockerPath);
@@ -77,7 +80,7 @@ class UpCommand extends AbstractCommand
 
         // Start current docker containers
         $this->output->writeln('<p>Start docker containers in "' . $dockerPath . '"</p>');
-        $command = new CommandBuilder(null, 'up -d');
+        $command = new CommandBuilder(null, $dockerComposeArgs);
         $ret     = $this->executeDockerCompose($command);
 
         // Store docker path in settings (last docker startup)