+++ src/app/CliTools/Console/Command/Sync/BackupCommand.php
@@ -4,6 +4,7 @@ namespace CliTools\Console\Command\Sync;
 
 /*
  * CliTools Command
+ * Copyright (C) 2016 WebDevOps.io
  * Copyright (C) 2015 Markus Blaschke <markus@familie-blaschke.net>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -32,7 +33,7 @@ class BackupCommand extends AbstractShareCommand
     {
         parent::configure();
 
-        $this->setName('sync:backup')
+        $this->setName('share:backup')
              ->setDescription('Backup files and database from share');
     }