+++ src/app/CliTools/Console/Command/Sync/RestoreCommand.php
@@ -4,6 +4,7 @@ namespace CliTools\Console\Command\Sync;
 
 /*
  * CliTools Command
+ * Copyright (C) 2016 WebDevOps.io
  * Copyright (C) 2015 Markus Blaschke <markus@familie-blaschke.net>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -30,7 +31,7 @@ class RestoreCommand extends AbstractShareCommand
     {
         parent::configure();
 
-        $this->setName('sync:restore')
+        $this->setName('share:restore')
              ->setDescription('Restore files and database from share');
     }